+++ src/main/java/cn/nukkit/raknet/protocol/EncapsulatedPacket.java
@@ -18,7 +18,7 @@ public class EncapsulatedPacket implements Cloneable {
     public Integer orderIndex = null;
     public Integer orderChannel = null;
     public Integer splitCount = null;
-    public Short splitID = null;
+    public Integer splitID = null;
     public Integer splitIndex = null;
     public byte[] buffer;
     public boolean needACK = false;
